refactor(cohesion): group csp-safe execution into src/background/csp-safe/
The structured CSP-safe executor, its expression parser, and their shared types
change together but sat flat in src/background/ under a csp-safe-* prefix. Move
them into a cohesive folder (common-closure) and drop the now-redundant prefix:

  csp-safe-executor.ts -> csp-safe/executor.ts
  csp-safe-parser.ts   -> csp-safe/parser.ts
  csp-safe-types.ts    -> csp-safe/types.ts

Only query-execution.ts consumes them (2 imports, rewired). parser.ts's internal
imports adjusted for the extra depth (../lib -> ../../lib, sibling ./types).
tsc auto-mirrors the new folder into extension/background/csp-safe/; stale
compiled csp-safe-*.js removed. Three test files repointed to the new compiled
paths. csp-safe feature index + tech-spec code_paths updated (docs contract).

Behavior-preserving. typecheck + compile-ts clean; csp-safe (96) and
query-execution-routing (2) suites pass.
