+++ docs/features/feature/csp-safe-execution/index.md
@@ -4,11 +4,11 @@ feature_id: feature-csp-safe-execution
 status: implemented
 feature_type: feature
 owners: []
-last_reviewed: 2026-07-05
+last_reviewed: 2026-07-25
 code_paths:
-  - src/background/csp-safe-types.ts
-  - src/background/csp-safe-parser.ts
-  - src/background/csp-safe-executor.ts
+  - src/background/csp-safe/types.ts
+  - src/background/csp-safe/parser.ts
+  - src/background/csp-safe/executor.ts
   - src/background/query-execution.ts
   - src/inject/execute-js.ts
 test_paths:
@@ -43,9 +43,9 @@ Tier 2 is the big win: content scripts in ISOLATED world are exempt from page CS
 
 ## Code and Tests
 
-- Types: `src/background/csp-safe-types.ts`
-- Parser: `src/background/csp-safe-parser.ts`
-- Executor: `src/background/csp-safe-executor.ts`
+- Types: `src/background/csp-safe/types.ts`
+- Parser: `src/background/csp-safe/parser.ts`
+- Executor: `src/background/csp-safe/executor.ts`
 - Integration: `src/background/query-execution.ts`
 - Tests: `extension/background/__tests__/query-execution-serialization.test.js`
