+++ extension/background/csp-safe/executor.js
@@ -209,4 +209,4 @@ export function cspSafeExecutor(command) {
         };
     }
 }
-//# sourceMappingURL=csp-safe-executor.js.map
+//# sourceMappingURL=executor.js.map
