+++ extension/background/csp-safe/types.js
@@ -2,4 +2,4 @@
 // These types represent JavaScript expressions as JSON-serializable data structures
 // that can be interpreted without eval() or new Function().
 export {};
-//# sourceMappingURL=csp-safe-types.js.map
+//# sourceMappingURL=types.js.map
