+++ src/background/csp-safe/executor.ts
