+++ src/background/csp-safe/parser.ts
@@ -28,8 +28,8 @@
  * as structured commands and are rejected with guidance to use DOM primitives.
  */
 
-import type { StructuredCommand, StructuredValue, StructuredStep, ParseResult } from './csp-safe-types.js'
-import { errorMessage } from '../lib/error-utils.js'
+import type { StructuredCommand, StructuredValue, StructuredStep, ParseResult } from './types.js'
+import { errorMessage } from '../../lib/error-utils.js'
 
 // Patterns that we reject outright before attempting to parse
 const REJECTED_PATTERNS: Array<{ pattern: RegExp; reason: string }> = [
