+++ src/background/csp-safe/types.ts
