+++ tests/extension/csp-safe-integration.test.js
@@ -7,8 +7,8 @@
 import { describe, test, beforeEach, mock } from 'node:test'
 import assert from 'node:assert'
 
-const { parseExpression } = await import('../../extension/background/csp-safe-parser.js')
-const { cspSafeExecutor } = await import('../../extension/background/csp-safe-executor.js')
+const { parseExpression } = await import('../../extension/background/csp-safe/parser.js')
+const { cspSafeExecutor } = await import('../../extension/background/csp-safe/executor.js')
 
 // =============================================================================
 // End-to-end: parse + execute
